+++ plugins/README.md
@@ -2,13 +2,13 @@
 
 Plugins allow developers to extend the functionality of Agent through the use of middleware.
 
-### Middleware Plugins
+### Interceptor Plugins
 
-Middleware can be added to agent to customize the request and/or response by implementing the [Middleware](./middleware/registry.go) interface.
-This interface needs to define a `Handler()` method that returns a standard net/http middleware handler based on [http.Handler](https://golang.org/pkg/net/http/#Handler) interface.
-The custom structure can also define a set of fields that can be configured via `config.yaml`.
+Interceptors can be added to agent to customize the request and/or response by implementing the [Interceptor](./interceptors/registry.go) interface.
+This interface needs to define a `Handler()` method that returns a standard net/http middleware handler based on [http.Handler](https://golang.org/pkg/net/http/#Handler).
+The custom structure can also include a set of fields that can be configured via `config.yaml`.
 
-### Example Middleware definition
+### Example Interceptor definition
 ```go
 package example
 
@@ -16,43 +16,43 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/optimizely/agent/pkg/middleware"
+	"github.com/optimizely/agent/plugins/interceptors"
 )
 
-type MiddlewarePlugin struct {
+type Example struct {
 	// set of configuration fields
 	RequestHeader  string
 	ResponseHeader string
 	ContextValue   string
 }
 
-func (p *MiddlewarePlugin) Handler() func(next http.Handler) http.Handler {
+func (i *Example) Handler() func(next http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			r.Header.Add("X-Example-Request", p.RequestHeader)
+			r.Header.Add("X-Example-Request", i.RequestHeader)
 
 			// Example adding context to the request path
-			ctx := context.WithValue(r.Context(), "example-context", p.ContextValue)
+			ctx := context.WithValue(r.Context(), "example-context", i.ContextValue)
 
 			// Continuing with the normal serving
 			next.ServeHTTP(w, r.WithContext(ctx))
 
 			// Modify the response in some way
-			w.Header().Add("X-Example-Response", p.ResponseHeader)
+			w.Header().Add("X-Example-Response", i.ResponseHeader)
 		})
 	}
 }
 
-// Register our middleware as "example".
+// Register our interceptor as "example".
 func init() {
-	middleware.Add("example", func() middleware.Plugin {
-		return &MiddlewarePlugin{}
+	interceptors.Add("example", func() interceptors.Interceptor {
+		return &Example{}
 	})
 }
 ```
 
-Once your plugin is written, you'll need to add the plugin as an anonymous import into [all.go](./middleware/all/all.go).
-This step is necessary to register the plugin with Agent. 
+To make the interceptor available, add the plugin as an anonymous import into [all.go](./interceptors/all/all.go).
+This will register the plugin with Agent. 
 ```go
 package all
